refactor(test): extract SpaceCreatorService, TaskCreatorService and TeamCreatorService
