+++ src/test/kotlin/org/rucca/cheese/api/RankTest.kt
@@ -9,8 +9,6 @@
 
 package org.rucca.cheese.api
 
-import java.time.LocalDateTime
-import kotlin.math.floor
 import org.json.JSONObject
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation
@@ -18,8 +16,9 @@ import org.junit.jupiter.api.Order
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestMethodOrder
-import org.rucca.cheese.common.helper.toEpochMilli
 import org.rucca.cheese.common.persistent.IdType
+import org.rucca.cheese.utils.SpaceCreatorService
+import org.rucca.cheese.utils.TaskCreatorService
 import org.rucca.cheese.utils.UserCreatorService
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -38,23 +37,18 @@ import org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPat
 @TestMethodOrder(OrderAnnotation::class)
 class RankTest
 @Autowired
-constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCreatorService) {
+constructor(
+    private val mockMvc: MockMvc,
+    private val userCreatorService: UserCreatorService,
+    private val spaceCreatorService: SpaceCreatorService,
+    private val taskCreatorService: TaskCreatorService,
+) {
     private val logger = LoggerFactory.getLogger(javaClass)
     lateinit var creator: UserCreatorService.CreateUserResponse
     lateinit var creatorToken: String
     lateinit var participant: UserCreatorService.CreateUserResponse
     lateinit var participantToken: String
-    private var spaceName = "Test Space (${floor(Math.random() * 10000000000).toLong()})"
-    private var spaceIntro = "This is a test space."
-    private var spaceDescription = "Description of space"
-    private var spaceAvatarId = userCreatorService.testAvatarId()
     private var spaceId: IdType = -1
-    private val taskName = "Test Task (${floor(Math.random() * 10000000000).toLong()})"
-    private val taskIntro = "This is a test task."
-    private val taskDescription = "Description of task"
-    private val taskDeadline = LocalDateTime.now().plusDays(7).toEpochMilli()
-    private val taskMembershipDeadline = LocalDateTime.now().plusMonths(1).toEpochMilli()
-    private val taskSubmissionSchema = listOf(Pair("Text Entry", "TEXT"))
     private var taskId: IdType = -1
     private var taskId2: IdType = -1
     private var taskId3: IdType = -1
@@ -62,93 +56,6 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
     private var submissionId: IdType = -1
     private var submissionId2: IdType = -1
     private var submissionId3: IdType = -1
-
-    fun createSpace(
-        creatorToken: String,
-        spaceName: String,
-        spaceIntro: String,
-        spaceDescription: String,
-        spaceAvatarId: IdType,
-    ): IdType {
-        val request =
-            MockMvcRequestBuilders.post("/spaces")
-                .header("Authorization", "Bearer $creatorToken")
-                .contentType("application/json")
-                .content(
-                    """
-                {
-                    "name": "$spaceName",
-                    "intro": "$spaceIntro",
-                    "description": "$spaceDescription",
-                    "avatarId": $spaceAvatarId,
-                    "enableRank": false,
-                    "announcements": "[]",
-                    "taskTemplates": "[]"
-                }
-            """
-                )
-        val response = mockMvc.perform(request).andExpect(MockMvcResultMatchers.status().isOk)
-        val spaceId =
-            JSONObject(response.andReturn().response.contentAsString)
-                .getJSONObject("data")
-                .getJSONObject("space")
-                .getLong("id")
-        logger.info("Created space: $spaceId")
-        return spaceId
-    }
-
-    fun createTask(
-        creatorToken: String,
-        name: String,
-        submitterType: String,
-        deadline: Long,
-        resubmittable: Boolean,
-        editable: Boolean,
-        intro: String,
-        description: String,
-        submissionSchema: List<Pair<String, String>>,
-        team: IdType?,
-        space: IdType?,
-        rank: Int?,
-    ): IdType {
-        val request =
-            MockMvcRequestBuilders.post("/tasks")
-                .header("Authorization", "Bearer $creatorToken")
-                .contentType("application/json")
-                .content(
-                    """
-                {
-                  "name": "$name",
-                  "submitterType": "$submitterType",
-                  "deadline": "$deadline",
-                  "resubmittable": $resubmittable,
-                  "editable": $editable,
-                  "intro": "$intro",
-                  "description": "$description",
-                  "submissionSchema": [
-                    ${
-                        submissionSchema
-                            .map { """
-                                {
-                                  "prompt": "${it.first}",
-                                  "type": "${it.second}"
-                                }
-                            """ }
-                            .joinToString(",\n")
-                    }
-                  ],
-                  "team": ${team?: "null"},
-                  "space": ${space?: "null"},
-                  "rank": ${rank?: "null"}
-                }
-            """
-                )
-        val response = mockMvc.perform(request).andExpect(MockMvcResultMatchers.status().isOk)
-        val json = JSONObject(response.andReturn().response.contentAsString)
-        val taskId = json.getJSONObject("data").getJSONObject("task").getLong("id")
-        logger.info("Created task: $taskId")
-        return taskId
-    }
 
     fun approveTask(taskId: IdType, token: String) {
         val request =
@@ -199,9 +106,7 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
             .andExpect(
-                MockMvcResultMatchers.jsonPath(
-                        "$.data.participants[?(@.member.id == $memberId)].approved"
-                    )
+                jsonPath("$.data.participants[?(@.member.id == $memberId)].approved")
                     .value("APPROVED")
             )
     }
@@ -234,78 +139,58 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         creatorToken = userCreatorService.login(creator.username, creator.password)
         participant = userCreatorService.createUser()
         participantToken = userCreatorService.login(participant.username, participant.password)
-        spaceId = createSpace(creatorToken, spaceName, spaceIntro, spaceDescription, spaceAvatarId)
+        spaceId = spaceCreatorService.createSpace(creatorToken)
         taskId =
-            createTask(
+            taskCreatorService.createTask(
                 creatorToken,
-                taskName,
-                "USER",
-                taskDeadline,
-                false,
-                false,
-                taskIntro,
-                taskDescription,
-                taskSubmissionSchema,
-                null,
-                spaceId,
-                1,
+                submitterType = "USER",
+                resubmittable = false,
+                editable = false,
+                team = null,
+                space = spaceId,
+                rank = 1,
             )
         approveTask(taskId, creatorToken)
         addParticipantUser(participantToken, taskId, participant.userId)
         approveTaskParticipant(creatorToken, taskId, participant.userId)
         submissionId = submitTaskUser(participantToken, taskId, participant.userId)
         taskId2 =
-            createTask(
+            taskCreatorService.createTask(
                 creatorToken,
-                taskName,
-                "USER",
-                taskDeadline,
-                false,
-                false,
-                taskIntro,
-                taskDescription,
-                taskSubmissionSchema,
-                null,
-                spaceId,
-                2,
+                submitterType = "USER",
+                resubmittable = false,
+                editable = false,
+                team = null,
+                space = spaceId,
+                rank = 2,
             )
         approveTask(taskId2, creatorToken)
         addParticipantUser(participantToken, taskId2, participant.userId)
         approveTaskParticipant(creatorToken, taskId2, participant.userId)
         submissionId2 = submitTaskUser(participantToken, taskId2, participant.userId)
         taskId3 =
-            createTask(
+            taskCreatorService.createTask(
                 creatorToken,
-                taskName,
-                "USER",
-                taskDeadline,
-                false,
-                false,
-                taskIntro,
-                taskDescription,
-                taskSubmissionSchema,
-                null,
-                spaceId,
-                1,
+                submitterType = "USER",
+                resubmittable = false,
+                editable = false,
+                team = null,
+                space = spaceId,
+                rank = 1,
             )
         approveTask(taskId3, creatorToken)
         addParticipantUser(participantToken, taskId3, participant.userId)
         approveTaskParticipant(creatorToken, taskId3, participant.userId)
         submissionId3 = submitTaskUser(participantToken, taskId3, participant.userId)
         taskId4 =
-            createTask(
+            taskCreatorService.createTask(
                 creatorToken,
-                taskName,
-                "USER",
-                taskDeadline,
-                false,
-                false,
-                taskIntro,
-                taskDescription,
-                taskSubmissionSchema,
-                null,
-                spaceId,
-                2,
+                submitterType = "USER",
+                resubmittable = false,
+                editable = false,
+                team = null,
+                space = spaceId,
+                rank = 2,
             )
         approveTask(taskId4, creatorToken)
     }
@@ -320,8 +205,8 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.id").value(spaceId))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.myRank").isEmpty)
+            .andExpect(jsonPath("$.data.space.id").value(spaceId))
+            .andExpect(jsonPath("$.data.space.myRank").isEmpty)
     }
 
     @Test
@@ -335,8 +220,8 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(requestBuilders)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.spaces[0].id").value(spaceId))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.spaces[0].myRank").isEmpty)
+            .andExpect(jsonPath("$.data.spaces[0].id").value(spaceId))
+            .andExpect(jsonPath("$.data.spaces[0].myRank").isEmpty)
     }
 
     @Test
@@ -350,7 +235,7 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.enableRank").value(true))
+            .andExpect(jsonPath("$.data.space.enableRank").value(true))
     }
 
     @Test
@@ -363,8 +248,8 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.id").value(spaceId))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.myRank").value(0))
+            .andExpect(jsonPath("$.data.space.id").value(spaceId))
+            .andExpect(jsonPath("$.data.space.myRank").value(0))
     }
 
     @Test
@@ -378,8 +263,8 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(requestBuilders)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.spaces[0].id").value(spaceId))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.spaces[0].myRank").value(0))
+            .andExpect(jsonPath("$.data.spaces[0].id").value(spaceId))
+            .andExpect(jsonPath("$.data.spaces[0].myRank").value(0))
     }
 
     @Test
@@ -398,11 +283,9 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isBadRequest)
-            .andExpect(
-                MockMvcResultMatchers.jsonPath("$.error.name").value("YourRankIsNotHighEnoughError")
-            )
-            .andExpect(MockMvcResultMatchers.jsonPath("$.error.data.yourRank").value(0))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.error.data.requiredRank").value(1))
+            .andExpect(jsonPath("$.error.name").value("YourRankIsNotHighEnoughError"))
+            .andExpect(jsonPath("$.error.data.yourRank").value(0))
+            .andExpect(jsonPath("$.error.data.requiredRank").value(1))
     }
 
     @Test
@@ -424,9 +307,7 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(
-                MockMvcResultMatchers.jsonPath("$.data.hasUpgradedParticipantRank").value("false")
-            )
+            .andExpect(jsonPath("$.data.hasUpgradedParticipantRank").value("false"))
     }
 
     @Test
@@ -439,8 +320,8 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.id").value(spaceId))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.myRank").value(0))
+            .andExpect(jsonPath("$.data.space.id").value(spaceId))
+            .andExpect(jsonPath("$.data.space.myRank").value(0))
     }
 
     @Test
@@ -488,8 +369,8 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.id").value(spaceId))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.myRank").value(1))
+            .andExpect(jsonPath("$.data.space.id").value(spaceId))
+            .andExpect(jsonPath("$.data.space.myRank").value(1))
     }
 
     @Test
@@ -511,9 +392,7 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(
-                MockMvcResultMatchers.jsonPath("$.data.hasUpgradedParticipantRank").value("true")
-            )
+            .andExpect(jsonPath("$.data.hasUpgradedParticipantRank").value("true"))
     }
 
     @Test
@@ -526,8 +405,8 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.id").value(spaceId))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.myRank").value(2))
+            .andExpect(jsonPath("$.data.space.id").value(spaceId))
+            .andExpect(jsonPath("$.data.space.myRank").value(2))
     }
 
     @Test
@@ -549,9 +428,7 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(
-                MockMvcResultMatchers.jsonPath("$.data.hasUpgradedParticipantRank").value("false")
-            )
+            .andExpect(jsonPath("$.data.hasUpgradedParticipantRank").value("false"))
     }
 
     @Test
@@ -564,7 +441,7 @@ constructor(private val mockMvc: MockMvc, private val userCreatorService: UserCr
         mockMvc
             .perform(request)
             .andExpect(MockMvcResultMatchers.status().isOk)
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.id").value(spaceId))
-            .andExpect(MockMvcResultMatchers.jsonPath("$.data.space.myRank").value(2))
+            .andExpect(jsonPath("$.data.space.id").value(spaceId))
+            .andExpect(jsonPath("$.data.space.myRank").value(2))
     }
 }
