+++ src/test/kotlin/org/rucca/cheese/api/TaskSubmissionReviewTest.kt
@@ -9,8 +9,6 @@
 
 package org.rucca.cheese.api
 
-import java.time.LocalDateTime
-import kotlin.math.floor
 import org.json.JSONObject
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation
@@ -18,9 +16,9 @@ import org.junit.jupiter.api.Order
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.TestMethodOrder
-import org.rucca.cheese.common.helper.toEpochMilli
 import org.rucca.cheese.common.persistent.IdType
 import org.rucca.cheese.utils.AttachmentCreatorService
+import org.rucca.cheese.utils.TaskCreatorService
 import org.rucca.cheese.utils.UserCreatorService
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -42,6 +40,7 @@ constructor(
     private val mockMvc: MockMvc,
     private val userCreatorService: UserCreatorService,
     private val attachmentCreatorService: AttachmentCreatorService,
+    private val taskCreatorService: TaskCreatorService,
 ) {
     private val logger = LoggerFactory.getLogger(javaClass)
     lateinit var creator: UserCreatorService.CreateUserResponse
@@ -51,62 +50,6 @@ constructor(
     private var attachmentId: IdType = -1
     private var taskId: IdType = -1
     private var submissionId: IdType = -1
-    private val taskName = "Test Task (${floor(Math.random() * 10000000000).toLong()})"
-    private val taskIntro = "This is a test task."
-    private val taskDescription = "A lengthy text. ".repeat(1000)
-    private val taskDeadline = LocalDateTime.now().plusDays(7).toEpochMilli()
-    private val taskSubmissionSchema =
-        listOf(Pair("Text Entry", "TEXT"), Pair("Attachment Entry", "FILE"))
-
-    fun createTask(
-        name: String,
-        submitterType: String,
-        deadline: Long,
-        resubmittable: Boolean,
-        editable: Boolean,
-        intro: String,
-        description: String,
-        submissionSchema: List<Pair<String, String>>,
-        team: IdType?,
-        space: IdType?,
-    ): IdType {
-        val request =
-            MockMvcRequestBuilders.post("/tasks")
-                .header("Authorization", "Bearer $creatorToken")
-                .contentType("application/json")
-                .content(
-                    """
-                {
-                  "name": "$name",
-                  "submitterType": "$submitterType",
-                  "deadline": "$deadline",
-                  "resubmittable": $resubmittable,
-                  "editable": $editable,
-                  "intro": "$intro",
-                  "description": "$description",
-                  "submissionSchema": [
-                    ${
-                        submissionSchema
-                            .map { """
-                                {
-                                  "prompt": "${it.first}",
-                                  "type": "${it.second}"
-                                }
-                            """ }
-                            .joinToString(",\n")
-                    }
-                  ],
-                  "team": ${team?: "null"},
-                  "space": ${space?: "null"}
-                }
-            """
-                )
-        val response = mockMvc.perform(request).andExpect(MockMvcResultMatchers.status().isOk)
-        val json = JSONObject(response.andReturn().response.contentAsString)
-        val taskId = json.getJSONObject("data").getJSONObject("task").getLong("id")
-        logger.info("Created task: $taskId")
-        return taskId
-    }
 
     fun joinTask(taskId: IdType, participantId: IdType, participantToken: String) {
         val request =
@@ -179,17 +122,10 @@ constructor(
         participantToken = userCreatorService.login(participant.username, participant.password)
         attachmentId = attachmentCreatorService.createAttachment(creatorToken)
         taskId =
-            createTask(
-                taskName,
-                "USER",
-                taskDeadline,
-                false,
-                false,
-                taskIntro,
-                taskDescription,
-                taskSubmissionSchema,
-                null,
-                null,
+            taskCreatorService.createTask(
+                creatorToken,
+                submissionSchema =
+                    listOf(Pair("Text Entry", "TEXT"), Pair("Attachment Entry", "FILE")),
             )
         joinTask(taskId, participant.userId, participantToken)
         approveTaskParticipant(creatorToken, taskId, participant.userId)
